placement-portal-react-fastapi/backend/requirements.txt

fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
passlib[bcrypt]==1.7.4
PyJWT==2.9.0
